Moved offset calc to utils/signals
